Fixes #485 by indenting the second line of text
This solution solves the problem by indenting the second line of text
(and beyond).  This is a pure-CSS solution, which should be preferrable
to a solution in JS.

This is not necessarily the requested solution, but it looks clean and
solves the problem in a simple way.

Tested other screen sizes and looks good in all sizes.